requirements.txt
werkzeug==0.16.1
flask-restplus==0.13.0
retrying==1.3.3
Flask-PyMongo==2.3.0
mongodb-migrations==0.7.0
flask-cors==3.0.8
